ddpdoj: fix D43 -- the laser bomb's box was signed, the cartridge's is unsigned
$2457A0, $2457A8, $2457B8 and $2457C0 are all 65 xx = bcs, unsigned
lower. A signed compare would be 6D. recordHitsBox wrote them as i16()
while quoting bcs in the comment beside each one, and the same slip
sits on the eight twins at $2458B6/$2458BE/$2458CE/$2458D6 and
$245990/$245998/$2459A8/$2459B0. Verified all four bytes myself.

Every coordinate there carries D6's $2800 bias, so a raw Y of $5800 or
more biases past $8000 and reads negative when signed. The boss's
biased far edge is $9F7D on every checkpoint of the fight, and the
beam's segments cross $8000 from segment 24 out.

Measured over 131 damage frames: intersecting segments found 0 to 10,
pool-B hits 0 to 64, boss HP $7FFF untouched to $FDFF. $7FFF minus 64
times $208 is $FDFF exactly.

My brief offered two options, "never intersects" or "loses the nearest
test", and the answer was neither: it intersected, nothing beat it,
and the port's own comparison rejected all ten segments so the nearest
test was never reached. Chasing "what won" would have burned the wave.

My pool order was also backwards and it is load-bearing: the ROM runs
pool B first and the pool-B target SHADOWS pool A, so anything behind
the nearest pool-B record takes nothing. And in a boss fight pool A is
0 of 100 live for the whole encounter, so the laser bomb's hit-
everything pass has nothing to hit. That part is the cartridge.

D6 and the $9800 reject were both fine, but there is a second
move.w #$2800,D6 at $24535A inside a movem-saved subroutine, which a
narrower reader would mistake for a rewrite.

The bullets share recordHitsBox and carried the same defect; their
erased count moved the opposite way, because the signed form was wrong
in both directions depending on which side of $8000 an edge fell.

No gate baseline moved. Suite 3748 pass / 0 fail / 0 skipped, gate
exit 0, 600 windows, verified on my own run.

+++ games/ddpdoj/src/bomb.js
@@ -1210,16 +1210,35 @@ function beamBox2456A6(ram) {
 /** `$245788..$2457EC`, `$24589E..$2458F4` and `$245978..$2459C0` -- the SAME
  *  AABB against ONE bomb record, at three call sites with three different
  *  starting records and counts.  The bullets' site passes `d0 === d1` and
- *  `d2 === d3`, i.e. a POINT, and that is the only difference. */
+ *  `d2 === d3`, i.e. a POINT, and that is the only difference.
+ *
+ *  **ALL FOUR TESTS ARE UNSIGNED, AND W413 FIXED THEM.**  `$2457A0`, `$2457A8`,
+ *  `$2457B8`, `$2457C0` are `65 xx` -- `bcs`, the CARRY, i.e. unsigned LOWER --
+ *  and so are their twins at `$2458B6`/`$2458BE`/`$2458CE`/`$2458D6` and
+ *  `$245990`/`$245998`/`$2459A8`/`$2459B0`.  The first draft wrote all twelve
+ *  as `i16(...)` signed compares while quoting `bcs` in the comment beside
+ *  them, which is the shape W411 warns about: a wrong reading with the right
+ *  citation next to it.
+ *
+ *  **IT IS NOT A CORNER CASE, IT IS THE TOP OF THE SCREEN.**  Every coordinate
+ *  here carries D6's `$2800` bias, so a raw Y of `$5800` and up biases to
+ *  `$8000` and up and reads NEGATIVE as `i16`.  `[M]` the stage-1 boss's own
+ *  sub-record, taken verbatim from the board's RAM at lf9,000
+ *  (`$81523C`: Y `$697D`, half-extents `$0E00`/`$1780`/`$0800`/`$0800`), has a
+ *  biased far edge of `$9F7D` on EVERY checkpoint of the fight, and the beam's
+ *  own segments bias past `$8000` from segment 24 outward.  Signed, ten
+ *  intersecting segments were rejected -- nine at test 1, one at test 2 -- and
+ *  `$812954` stayed 0, so `$2457FA tst.w / beq` skipped the ONE pool-B hit and
+ *  the boss took nothing for all 131 damage frames.  Unsigned, it takes it. */
 function recordHitsBox(ram, a6, d6, d0, d1, d2, d3) {
-  const d4y = i16(u16(u16(ram.u16(a6 + 0x02) + d6) + ram.u16(a6 + 0x10)));
-  if (d4y < i16(d1)) return false;                     // $24579E cmp.w D1,D4/bcs
-  const d5y = i16(u16(u16(ram.u16(a6 + 0x02) + d6) - ram.u16(a6 + 0x12)));
-  if (i16(d0) < d5y) return false;                     // $2457A6 cmp.w D5,D0/bcs
-  const d4x = i16(u16(u16(ram.u16(a6 + 0x04) + d6) + ram.u16(a6 + 0x14)));
-  if (d4x < i16(d3)) return false;                     // $2457B6 cmp.w D3,D4/bcs
-  const d5x = i16(u16(u16(ram.u16(a6 + 0x04) + d6) - ram.u16(a6 + 0x16)));
-  if (i16(d2) < d5x) return false;                     // $2457BE cmp.w D5,D2/bcs
+  const d4y = u16(u16(ram.u16(a6 + 0x02) + d6) + ram.u16(a6 + 0x10));
+  if (d4y < d1) return false;                          // $24579E cmp.w D1,D4/bcs
+  const d5y = u16(u16(ram.u16(a6 + 0x02) + d6) - ram.u16(a6 + 0x12));
+  if (d0 < d5y) return false;                          // $2457A6 cmp.w D5,D0/bcs
+  const d4x = u16(u16(ram.u16(a6 + 0x04) + d6) + ram.u16(a6 + 0x14));
+  if (d4x < d3) return false;                          // $2457B6 cmp.w D3,D4/bcs
+  const d5x = u16(u16(ram.u16(a6 + 0x04) + d6) - ram.u16(a6 + 0x16));
+  if (d2 < d5x) return false;                          // $2457BE cmp.w D5,D2/bcs
   return true;
 }
